Move NotificationModal into components/modals folder
