+++ src/components/modals/index.ts
@@ -1,1 +1,2 @@
 export { ModalPortal } from './ModalPortal'
+export * from './NotificationModal'
